Use FRMS6 for additional backend test since HDF5 rejects alternative IO backends
