+++ tests/utils.py
@@ -14,6 +14,7 @@ from libertem.udf.raw import PickUDF
 from libertem.udf.masks import ApplyMasksUDF
 from libertem.corrections import CorrectionSet
 from libertem.corrections.detector import correct
+from libertem.io.dataset.base.backend import IOBackend, IOBackendImpl
 
 
 def _naive_mask_apply(masks, data):
@@ -242,3 +243,19 @@ def get_testdata_path():
             os.path.join(os.path.dirname(__file__), '..', 'data')
         )
     )
+
+
+class FakeBackend(IOBackend, id_="fake"):
+    def get_impl(self):
+        return FakeBackendImpl()
+
+
+class FakeBackendImpl(IOBackendImpl):
+    def get_tiles(
+        self, tiling_scheme, fileset, read_ranges, roi, native_dtype, read_dtype, decoder,
+        sync_offset, corrections,
+    ):
+        raise RuntimeError("nothing to see here")
+        # to make this a generator, there needs to be a yield statement in
+        # the body of the function, even if it is never executed:
+        yield
